zip2_extracted/iracing-enduro-client/requirements.txt
pyirsdk==1.3.2
requests==2.31.0
pystray==0.19.5
Pillow==10.3.0
packaging==24.0
